requirements.txt
streamlit==1.39.0
google-generativeai>=0.7.0
pandas==2.2.2
numpy==2.1.1
python-dotenv==1.0.1
protobuf>=4.21.0
requests
